refactored variance functions
