+++ pyproject.toml
@@ -113,8 +113,8 @@ lint = [
   "mypy src/",
 ]
 format = ["black .", "isort .", "lint",]
-test = ["pytest -n auto ./src/ ./tests/",]
-fast-test = ["pytest -n auto ./tests/ -m \"not slow\"",]
+test = ["pytest -n auto --dist worksteal ./src/ ./tests/",]
+fast-test = ["pytest -n auto --dist worksteal ./tests/ -m \"not slow\"",]
 coverage = [
   "pytest --cov --cov-report term --cov-report xml --junitxml=xunit-result.xml tests/",
 ]
@@ -150,7 +150,7 @@ select = [  # It would be nice to have the commented out checks working.
   # "D",  # pydocstyle
   # "C90",  # mccabe complexity
   # "N",  # PEP8-naming
-  "UP",  # pyupgrade (upgrade syntax to current syntax)
+  # "UP",  # pyupgrade (upgrade syntax to current syntax)
   "PLE",  # Pylint error https://github.com/charliermarsh/ruff#error-ple
   # "PLR",  # Pylint refactor (e.g. too-many-arguments)
   # "PLW",  # Pylint warning (useless-else-on-loop)
