+++ src/dtscalibration/calibrate_utils.py
@@ -1,4 +1,3 @@
-# coding=utf-8
 import numpy as np
 import scipy.sparse as sp
 import statsmodels.api as sm
